// interfaces/menu.interfaces.ts
import { TopPageCategory } from "./top-page.interface";

export interface PageItem {
    alias: string,
    title: string,
    _id: string,
    category: string
}

export interface MenuItem {
    _id: {
        secondCategory: string
    },
    // для третьего уровня меню
    isOpened?: boolean;
    pages: PageItem[];
}

export interface FirstLevelMenuItem {
    route: string;
    name: string;
    icon: JSX.Element;
    id: TopPageCategory;
}